feat: upload stream functionality added
- UploadStream function upload by chunking of size 4MB and with 5 go routines
- This function is used when file size is bigger then 32MB
- These magic numbers are based on fog/azurebs library, check in this link  https://github.com/fog/fog-azure-rm/blob/68ea3f7dd14c0a524146f1e426fa2ad1f6192cc8/lib/fog/azurerm/constants.rb

+++ azurebs/client/client.go
@@ -15,6 +15,17 @@ type AzBlobstore struct {
 	storageClient StorageClient
 }
 
+// Single blob put threshold is 32MB
+const singleBlobPutThreshold = int64(32 * 1024 * 1024)
+
+func getFileSize(source *os.File) (int64, error) {
+	fileInfo, err := source.Stat()
+	if err != nil {
+		return 0, fmt.Errorf("failed to get file stat: %w", err)
+	}
+	return fileInfo.Size(), nil
+}
+
 func New(storageClient StorageClient) (AzBlobstore, error) {
 	return AzBlobstore{storageClient: storageClient}, nil
 }
@@ -30,24 +41,36 @@ func (client *AzBlobstore) Put(sourceFilePath string, dest string) error {
 		return err
 	}
 	defer source.Close() //nolint:errcheck
-
-	md5, err := client.storageClient.Upload(source, dest)
+	fileSize, err := getFileSize(source)
 	if err != nil {
-		return fmt.Errorf("upload failure: %w", err)
+		return err
 	}
-
-	if !bytes.Equal(sourceMD5, md5) {
-		slog.Error("Upload failed due to MD5 mismatch, deleting blob", "blob", dest, "expected_md5", fmt.Sprintf("%x", sourceMD5), "received_md5", fmt.Sprintf("%x", md5))
-
-		err := client.storageClient.Delete(dest)
+	if fileSize <= singleBlobPutThreshold {
+		md5, err := client.storageClient.Upload(source, dest)
 		if err != nil {
-			slog.Error("Failed to delete blob after MD5 mismatch", "blob", dest, "error", err)
-
+			return fmt.Errorf("upload failure: %w", err)
 		}
-		return fmt.Errorf("MD5 mismatch: expected %x, got %x", sourceMD5, md5)
+
+		if !bytes.Equal(sourceMD5, md5) {
+			slog.Error("Upload failed due to MD5 mismatch, deleting blob", "blob", dest, "expected_md5", fmt.Sprintf("%x", sourceMD5), "received_md5", fmt.Sprintf("%x", md5))
+
+			err := client.storageClient.Delete(dest)
+			if err != nil {
+				slog.Error("Failed to delete blob after MD5 mismatch", "blob", dest, "error", err)
+
+			}
+			return fmt.Errorf("MD5 mismatch: expected %x, got %x", sourceMD5, md5)
+		}
+
+		slog.Debug("MD5 verification passed", "blob", dest, "md5", fmt.Sprintf("%x", md5))
+
+	} else {
+		err := client.storageClient.UploadStream(source, dest)
+		if err != nil {
+			return fmt.Errorf("upload failure: %w", err)
+		}
 	}
 
-	slog.Debug("MD5 verification passed", "blob", dest, "md5", fmt.Sprintf("%x", md5))
 	return nil
 }
